+++ src/datafactory/azext_datafactory/vendored_sdks/datafactory/aio/_configuration.py
@@ -12,11 +12,12 @@ from azure.core.configuration import Configuration
 from azure.core.pipeline import policies
 from azure.mgmt.core.policies import ARMHttpLoggingPolicy
 
+from .._version import VERSION
+
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
     from azure.core.credentials_async import AsyncTokenCredential
 
-VERSION = "unknown"
 
 class DataFactoryManagementClientConfiguration(Configuration):
     """Configuration for DataFactoryManagementClient.
@@ -46,7 +47,7 @@ class DataFactoryManagementClientConfiguration(Configuration):
         self.subscription_id = subscription_id
         self.api_version = "2018-06-01"
         self.credential_scopes = kwargs.pop('credential_scopes', ['https://management.azure.com/.default'])
-        kwargs.setdefault('sdk_moniker', 'datafactorymanagementclient/{}'.format(VERSION))
+        kwargs.setdefault('sdk_moniker', 'mgmt-datafactory/{}'.format(VERSION))
         self._configure(**kwargs)
 
     def _configure(