+++ src/ssh/azext_ssh/tests/latest/test_custom.py
@@ -5,7 +5,7 @@
 
 import io
 from azure.cli.core import azclierror
-import mock
+from unittest import mock
 import unittest
 
 from azext_ssh import custom