+++ src/virtual-wan/azext_vwan/custom.py
@@ -153,11 +153,11 @@ def create_virtual_hub(cmd, resource_group_name, virtual_hub_name, address_prefi
 def get_effective_virtual_hub_routes(cmd, resource_group_name, virtual_hub_name,
                                      virtual_wan_resource_type=None, resource_id=None, no_wait=False):
     parameters = None
-    Resource, EffectiveRoutesParameters = cmd.get_models("Resource", 'EffectiveRoutesParameters')
+    EffectiveRoutesParameters = cmd.get_models('EffectiveRoutesParameters')
     if virtual_wan_resource_type is not None or resource_id is not None:
         parameters = EffectiveRoutesParameters(
             virtual_wan_resource_type=virtual_wan_resource_type,
-            resource_id=Resource(id=resource_id)
+            resource_id=resource_id
         )
 
     client = network_client_factory(cmd.cli_ctx).virtual_hubs
@@ -588,6 +588,9 @@ def add_vpn_gateway_connection_ipsec_policy(cmd, resource_group_name, gateway_na
     client = network_client_factory(cmd.cli_ctx).vpn_gateways
     gateway = client.get(resource_group_name, gateway_name)
     conn = _find_item_at_path(gateway, 'connections.{}'.format(connection_name))
+
+    if conn.ipsec_policies is None:
+        conn.ipsec_policies = []
     conn.ipsec_policies.append(
         IpsecPolicy(
             sa_life_time_seconds=sa_life_time_seconds,
@@ -600,6 +603,7 @@ def add_vpn_gateway_connection_ipsec_policy(cmd, resource_group_name, gateway_na
             pfs_group=pfs_group
         )
     )
+
     _upsert(gateway, 'connections', conn, 'name', warn=False)
     poller = sdk_no_wait(no_wait, client.create_or_update,
                          resource_group_name, gateway_name, gateway)